Part of the same move.

+++ commons/httpfetch/httpfetch.h
@@ -81,11 +81,11 @@ namespace Fetch {
 		
 	public:
 		HTTPFetch(string&);
-		HTTPFetch(const string&,const string&,const string&,string&,int,int,string&);
+		HTTPFetch(const string&,const string&,const string&,string&,unsigned long long,unsigned long long,string&);
 		~HTTPFetch();
 //		bool fetchPage(std::string, HTTPFetchHeader&, Redirects&, std::string&,int,int, std::string&);
 		void addHeader(std::string&);
-		bool doRequest(string&,string&,int, int, string&);
+		bool doRequest(string&,string&,unsigned long long,unsigned long long, string&);
 		static bool startup();
 		static bool available();
 		static bool shutdown();
